Update to expand grouped workout blocks
Biggest change is the ability to see avg power for "grouped" workout blocks that have "On" and "Off" intervals that repeat some number of times. The average power metric can't be calculated individually for each of these repeat blocks, so now the script breaks them up into individual intervals when it comes across those so average power and be correctly computed and displayed. 

Also made it create a new folder with the modified files. 

Also added some error handling